Move code around to turn some stages into packages
