+++ local-docker/entrypoint.py
@@ -53,8 +53,8 @@ from subprocess import check_output, Popen
 from socket import gethostbyname, gaierror
 from time import time, sleep
 
-from telepresence.expose import expose_local_services
-from telepresence.ssh import SSH
+from telepresence.connect.expose import expose_local_services
+from telepresence.connect.ssh import SSH
 from telepresence.runner import Runner, wait_for_exit
