+++ telepresence/connect/__init__.py
@@ -16,14 +16,12 @@ import argparse
 import re
 from typing import Tuple
 
-from telepresence.background import launch_local_server
-from telepresence.container import MAC_LOOPBACK_IP
-from telepresence.deployment import create_new_deployment, \
-    swap_deployment_openshift, supplant_deployment
-from telepresence.expose import expose_local_services
-from telepresence.remote import RemoteInfo, get_remote_info
+from telepresence.runner.background import launch_local_server
+from telepresence.connect.expose import expose_local_services
+from telepresence.connect.ssh import SSH
+from telepresence.startup import MAC_LOOPBACK_IP
+from telepresence.proxy.remote import RemoteInfo
 from telepresence.runner import Runner
-from telepresence.ssh import SSH
 from telepresence.utilities import find_free_port
 
 
@@ -149,76 +147,3 @@ def connect(
 
     span.end()
     return socks_port, ssh
-
-
-def start_proxy(runner: Runner, args: argparse.Namespace) -> RemoteInfo:
-    """Start the kubectl port-forward and SSH clients that do the proxying."""
-    span = runner.span()
-    if runner.chatty and args.method != "container":
-        runner.show(
-            "Starting proxy with method '{}', which has the following "
-            "limitations:".format(args.method)
-        )
-        if args.method == "vpn-tcp":
-            runner.show(
-                "All processes are affected, only one telepresence"
-                " can run per machine, and you can't use other VPNs."
-                " You may need to add cloud hosts with --also-proxy."
-            )
-        elif args.method == "inject-tcp":
-            runner.show(
-                "Go programs, static binaries, suid programs, and custom DNS"
-                " implementations are not supported."
-            )
-        runner.show(
-            "For a full list of method limitations see "
-            "https://telepresence.io/reference/methods.html\n"
-        )
-    if args.mount and runner.chatty:
-        runner.show(
-            "\nVolumes are rooted at $TELEPRESENCE_ROOT. See "
-            "https://telepresence.io/howto/volumes.html for details."
-        )
-
-    run_id = None
-
-    if args.new_deployment is not None:
-        # This implies --new-deployment:
-        args.deployment, run_id = create_new_deployment(runner, args)
-
-    if args.swap_deployment is not None:
-        # This implies --swap-deployment
-        if runner.kubectl.command == "oc":
-            args.deployment, run_id, container_json = (
-                swap_deployment_openshift(runner, args)
-            )
-        else:
-            args.deployment, run_id, container_json = supplant_deployment(
-                runner, args
-            )
-        args.expose.merge_automatic_ports([
-            p["containerPort"] for p in container_json.get("ports", [])
-            if p["protocol"] == "TCP"
-        ])
-
-    deployment_type = "deployment"
-    if runner.kubectl.command == "oc":
-        # OpenShift Origin uses DeploymentConfig instead, but for swapping we
-        # mess with ReplicationController instead because mutating DC doesn't
-        # work:
-        if args.swap_deployment:
-            deployment_type = "rc"
-        else:
-            deployment_type = "deploymentconfig"
-
-    remote_info = get_remote_info(
-        runner,
-        args.deployment,
-        args.context,
-        args.namespace,
-        deployment_type,
-        run_id=run_id,
-    )
-    span.end()
-
-    return remote_info
